Update neural_sim_main.py
I guess you should initialize the objector detector likewise the epoch=0. Otherwise, it would be hard to tell if the high accuracy is the impact of your proposed method or over-training of the object detector. Please clarify. Thank you in advance. I will raise an issue to have more discussion about my findings.